refactor: move ScoredMemoryEntry into its own module
