+++ src/models.rs
@@ -17,8 +17,10 @@ pub mod memory {
     }
 
     mod memory_entry;
+    mod scored_memory_entry;
     pub use enums::memory_type::MemoryType;
-    pub(crate) use memory_entry::{MemoryEntry, ScoredMemoryEntry};
+    pub(crate) use memory_entry::MemoryEntry;
+    pub(crate) use scored_memory_entry::ScoredMemoryEntry;
 }
 
 pub(crate) mod vector {
